Refactor: Remove Matchmaking API Scope
- Put Matchmaking functions into Rooms module
- serverlist_updater now not a child of Matchmaking scope
